# generated/openapi/fintech.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "proto/payments/payments.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/flows": {
      "get": {
        "operationId": "FlowService_ListFlows",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/flowListFlowsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "zoneId",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "FlowService"
        ]
      },
      "post": {
        "operationId": "FlowService_CreateFlow",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/flowFlow"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/flowCreateFlowRequest"
            }
          }
        ],
        "tags": [
          "FlowService"
        ]
      }
    },
    "/v1/flows/bulk-update": {
      "post": {
        "operationId": "FlowService_BulkUpdateFlows",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/flowBulkUpdateFlowsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/flowBulkUpdateFlowsRequest"
            }
          }
        ],
        "tags": [
          "FlowService"
        ]
      }
    },
...
    "/v1/zones": {
      "get": {
        "operationId": "ZoneService_ListZones",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/zoneListZonesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "orgId",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "ZoneService"
        ]
      },
      "post": {
        "operationId": "ZoneService_CreateZone",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/zoneZone"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/zoneCreateZoneRequest"
            }
          }
        ],
        "tags": [
          "ZoneService"
        ]
      }
    },
    "/v1/zones/bulk-metadata": {
      "post": {
        "operationId": "ZoneService_BulkUpdateMetadata",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/zoneBulkUpdateMetadataResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/zoneBulkUpdateMetadataRequest"
            }
          }
        ],
        "tags": [
          "ZoneService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "flowBulkUpdateFlowsRequest": {
      "type": "object",
      "properties": {
        "flowIds": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "enabled": {
          "type": "boolean"
        }
      }
    },
    "flowBulkUpdateFlowsResponse": {
      "type": "object",
      "properties": {
        "updatedCount": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "flowCreateFlowRequest": {
      "type": "object",
      "properties": {
        "zoneId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "nodesJson": {
          "type": "string"
        },
        "edgesJson": {
          "type": "string"
        }
      }
    },
...
    "flowFlow": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "zoneId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "enabled": {
          "type": "boolean"
        },
        "nodesJson": {
          "type": "string",
          "title": "Simplified for dynamic nodes"
        },
        "edgesJson": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "flowListFlowsResponse": {
      "type": "object",
      "properties": {
        "flows": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/flowFlow"
          }
        }
      }
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "zoneBulkUpdateMetadataRequest": {
      "type": "object",
      "properties": {
        "zoneIds": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "metadata": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      }
    },
    "zoneBulkUpdateMetadataResponse": {
      "type": "object",
      "properties": {
        "updatedCount": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "zoneCreateZoneRequest": {
      "type": "object",
      "properties": {
        "orgId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "mode": {
          "type": "string"
        },
        "templateName": {
          "type": "string"
        },
        "metadata": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      }
    },
    "zoneListZonesResponse": {
      "type": "object",
      "properties": {
        "zones": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/zoneZone"
          }
        }
      }
    },
    "zoneZone": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "orgId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "mode": {
          "type": "string",
          "title": "\"test\" | \"live\""
        },
        "secretKey": {
          "type": "string"
        },
        "publishableKey": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "templateName": {
          "type": "string"
        },
        "metadata": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "createdAt": {
          "type": "string"
        }
      }
    }
  }
}